[Refactor]: Decompose Bootstrapper into separate classes
This commit attempts to refactor the bootstrapper
class by breaking it into separate classes.

We create a ResolverManager class which handles
certain functionality needed by bootstrap.

The new class:

Uses resolver.py as a building block (calls resolver.resolve(), resolver.GenericProvider, etc.)
- Adds caching - avoids redundant resolution via _resolved_requirements dict
- Adds history-based resolution - reuses versions from previous bootstrap runs via prev_graph
- Handles git URL cloning - clones repos and extracts versions from package metadata
- Handles cached wheel lookups - checks local filesystem + remote cache server

This refactoring was suggested as a good first step
for multiple version bootstrap and also makes the code
simpler.

Fixes: https://github.com/python-wheel-build/fromager/issues/911
